requirements.txt
google-genai
pillow
ipython
streamlit
python-dotenv